Re-org to move ui into it's own package and add username fetch
